+++ apps/user/src/hooks/useShowSelections.ts
@@ -1,8 +1,8 @@
 import { useState } from "react";
 
-export const useShowSelection = (initialDate = "", initialName = "") => {
-  const [selectedShowDate, setSelectedShowDate] = useState(initialDate);
-  const [selectedShowName, setSelectedShowName] = useState(initialName);
+export const useShowSelection = () => {
+  const [selectedShowDate, setSelectedShowDate] = useState("");
+  const [selectedShowName, setSelectedShowName] = useState("");
 
   return {
     selectedShowDate,
